+++ assets/src/data/model/registration/constants.js
@@ -1,0 +1,1 @@
+export const MODEL_NAME = 'registration';
